removed the connection strings for safety
