+++ src/main/java/task2/DataSourceConfig.java
@@ -1,0 +1,21 @@
+package task2;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class DataSourceConfig {
+
+    @Bean
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName("com.mysql.cj.jdbc.Driver");
+        dataSource.setUrl("jdbc:mysql://localhost:3306/courses");
+        dataSource.setUsername("root");
+        dataSource.setPassword("@MindGame@28");
+        return dataSource;
+    }
+}
